エルコイン、欠席情報双方の登録削除制御追加

+++ app/Http/Controllers/AbsenceController.php
@@ -66,6 +66,11 @@ class AbsenceController extends Controller
             $id = $request->id;
             $item = Absence::find($id);
             $mode = 'edit';
+
+            if($item->HurikaeStatus == 5){
+                //エルコイン変換済みのレコードは編集できない
+                return redirect()->back()->withInput()->withErrors(['alertErr'=>'エルコイン変換済みのため編集できません。']);
+            }
         }
 
         //ドロップダウンリスト用データ取得（#TODOキャッシュにしたい）
@@ -111,6 +116,12 @@ class AbsenceController extends Controller
         //フォームにない項目をセット
         $absence->setUpdateColumn();
 
+        //振替ステータスのセット
+        if(!$absence->setHurikaeStatus()){
+            //想定外の更新が行われようとしているときは、エラーとして戻す
+            return redirect()->back()->withInput()->withErrors(['alertErr'=>'振替ステータス更新ルールエラー']);
+        }
+
         //INSERT処理
         $absence->save();
         
@@ -118,21 +129,30 @@ class AbsenceController extends Controller
         $args = [
         ];
 
-        return redirect()->route('absenceList',$args)->with('alertComp',MessageConst::ADD_COMPLETED);
+        return redirect()->route('absenceList-sp',$args)->with('alertComp',MessageConst::ADD_COMPLETED);
     }
     // \absence\edit
     public function edit(AbsenceRequest $request){
         
-        //クエリ文字列にidがついていない場合はエラー
-        if(isset($request->id)){}
-
         $absence = Absence::find($request->id);
+
+        if(empty($absence)){
+            //何らかの理由で取得できない場合エラー
+            return redirect()->back()->withInput()->withErrors(['alertErr'=>'該当のレコードがありません。']);
+        }
+
         $form = $request->all();
         unset($form['_token']);
         $absence->fill($form);
 
         //フォームにない項目をセット
         $absence->setUpdateColumn();
+
+        //振替ステータスのセット
+        if(!$absence->setHurikaeStatus()){
+            //想定外の更新が行われようとしているときは、エラーとして戻す
+            return redirect()->back()->withInput()->withErrors(['alertErr'=>'振替ステータス更新ルールエラー']);
+        }
 
         //UPDATE処理
         $absence->save();
@@ -141,17 +161,23 @@ class AbsenceController extends Controller
         $args = [
         ];
 
-        return redirect()->route('absenceList',$args)->with('alertComp',MessageConst::EDIT_COMPLETED);
+        return redirect()->route('absenceList-sp',$args)->with('alertComp',MessageConst::EDIT_COMPLETED);
     }
     // \absence\delete
     public function delete(AbsenceRequest $request){
 
-        #TODO 振替実績日が入っている場合はエラーにする。
-        
-        //クエリ文字列にidがついていない場合はエラー
-        if(isset($request->id)){}
-
         $absence = Absence::find($request->id);
+
+        if(empty($absence)){
+            //何らかの理由で取得できない場合エラー
+            return redirect()->back()->withInput()->withErrors(['alertErr'=>'該当のレコードがありません。']);
+        }
+
+        #振替実績日が入っている場合、エルコインに変換済みの場合、期限切れの場合エラーとする
+        if($absence->HurikaeStatus != 0){
+            //何らかの理由で取得できない場合エラー
+            return redirect()->back()->withInput()->withErrors(['alertErr'=>'振替ステータスが'.$absence->HurikaeStatusName.'です。未振替以外のレコードは削除できません。']);
+        }
 
         //DELETE処理
         $absence->delete();
@@ -160,7 +186,7 @@ class AbsenceController extends Controller
         $args = [
         ];
 
-        return redirect()->route('absenceList',$args)->with('alertComp',MessageConst::DELETE_COMPLETED);
+        return redirect()->route('absenceList-sp',$args)->with('alertComp',MessageConst::DELETE_COMPLETED);
     }
     
 }
